Move utils into _internal
